modules/terraform-google-ip-range-datasource/scripts/requirements.txt
dnspython==1.16.0